Post-merge adjustments
